optimizing packet

+++ src/main/kotlin/me/fzzyhmstrs/amethyst_imbuement/screen/ImbuingTableScreenHandler.kt
@@ -740,14 +740,14 @@ class ImbuingTableScreenHandler(
         override fun bufClassWriter(buf: PacketByteBuf) {
             buf.writeShort(power)
             buf.writeShort(id)
-            buf.writeShort(level)
+            buf.writeByte(level)
             buf.writeByte(slot)
         }
 
         override fun bufClassReader(world: World, buf: PacketByteBuf): TableResult {
             val pow = buf.readShort().toInt()
             val i = buf.readShort().toInt()
-            val lvl = buf.readShort().toInt()
+            val lvl = buf.readByte().toInt()
             val slt = buf.readByte().toInt()
             return EnchantingResult(pow,i,lvl, slt)
         }
@@ -1156,7 +1156,7 @@ class ImbuingTableScreenHandler(
 
     }
 
-    class LevelLowErrorResult(private val augment: String,pow: Int): ErrorResult(listOf(Text.translatable("container.imbuing_table.level_low").formatted(Formatting.RED)),
+    class LevelLowErrorResult(private val augment: String, pow: Int): ErrorResult(listOf(Text.translatable("container.imbuing_table.level_low").formatted(Formatting.RED)),
         Text.literal(augment),
         Text.literal(augment).fillStyle(Style.EMPTY.withFont(Identifier("minecraft", "default"))),
         pow,
